Admin fixes, view fixes, artist creation
Added admin checkbox to user registration form.  Added Add Artist link
to control panel.  Adding albums works, but they still will not display
on the Browse All page yet.  I'm pretty sure the ShoppingCart is already
working but can't test until the albums are able to be displayed to be
added into the cart.  Working on it.