Changing Criteria commands in StoreController
Getting MethodMissing exceptions for both withCriteria and
createCriteria.  The syntax is correct; I've checked it tons of times
now against the Grails documentation.  I'm not sure what is missing that
these functions are apparently not being created for the Album class.

I keep adding flush:true to save() calls in an attempt to get things to
actually persist.  I have a feeling now that something is not configured
somewhere, perhaps DataSource.groovy